backend/requirements.txt
fastapi
uvicorn
SQLAlchemy>=1.4
PyMySQL
passlib[bcrypt]
python-jose[cryptography]
pydantic
python-dotenv
